Refactored Sample and Speed classes.
SampleAndControl class now has a single static member function for the main
control loop, it can only be started if the file system is ready, and it now
assumes that the write thread is ready and waiting to write data.  The shell
command is now a static member function of this class.

SampleBuffer now controls the writing of the data and must be initialized with
an open file handle in order to work properly.

SpeedController now enables and disables the motor enable line properly.  The
shell command to enable and disable the speed controller is now a static member
function of this class.

Need to verify configuration of TIM4 and associated GPIO's for speed
measurements.