PI yaw rate loop design and implementation complete.
This commit introduces a dependency on Matlab in order to tune the proportional
and integral gains that are used for tracking a yaw rate reference.

The design now has a LQR/LQG controller which stabilizes the bicycle; this is
the "inner loop".  The outer loop is the yaw rate PI control.

The controller gains are now static data of the YawRateController class.

Also implemented in this commit are settable speed thresholds to control at
what speed the estimator gets turned on and what speed the controller gets
turned on.  The estimation threshold should be at a rear wheel rate (forward
speed) that is greater (lower) than that of the control threshold.

Also settable is whether the outer PI loop is enabled.  If disabled, only the
LQR/LQG control will be used so set point tracking of yaw rate will have no
effect.