CMake build causes gdb segfault.  Pulling out.
There is something subtley different between the bin file generated by the
ChibiOS provide Makefile and the current CMake build that results in the binary
having the same exact binary size but for some reason the CMake version causes
a segfault on Luke's machine for both 2012 and 2013 toolchains.  On Oliver's
machine, it only segfaulted with the 2013 toolchain.

Moving on to more important things and just going to use the standard Makefile.