Add GPIOF_HW_SWITCH_PIN to board.h.
This pin on the MCU is connected to a switch to ground. The pull-up resistors
on the pin make the default pin state high, when the button is pressed the pin
state will be low. The switch should be depressed while the bike is being
balanced by the experimental assistant, and released at the same time the
assistant is no longer touching the bike.