Change control threshold speed to delay by N samples.
Previous to this commit, there were two threshold speeds that needed to be
specified: estimation threshold speed and control threshold speed. After this
commit, control will turn on a fixed number of samples after estimation has
begun. The default for this is 10 samples, or 50ms.  The rationale for this is
that our estimator converges so rapidly that we can basically think of just one
speed now: the speed at which we want estimation and control to begin.