Fix static initialization bug that screwed steer angle calibration.
A namespace scope int32_t which was assigned the value of a constexpr for some
reason was not being initialized to the value of the constexpr. The result was
that the steer angle offset was being set to zero when the homefork function
was run, rather than set to the correct offset. This resulted in the "zero"
steer to be off by about 7 degrees.