Change speed control strategy.
Before we reach the rear wheel rate set point (a negative number for a forward
moving bicycle), we do apply maximum current. Once we have reach the set point,
the error will become >= 0.0, and we switch to PI control. We should
observe the bicycle reach the set point as rapidly as we can hope, and then
oscillate slightly about the set point but slowly converge to the set point.
The integral time Ti is set to 100.0s, so these oscillation should be very
minor and convergence quite slow.