Fix code that sets x acceleration threshold.
Weird bugs occurred with static data not getting initialized. Further, when
argc == 0, for some reason argc != 1 was evaluating to false and input
validation code was not being run.  Inspection with gdb confirmed the value of
the input parameter argument, but the if statement was not evaluating to true.
Weirdest bug ever.  Witnessed by Oliver, I am not crazy.